fix: print trace and abort message by default (#78)
When `env.trace` and `env.abort` are not provided by configuration. The
old behavior will try to inject a empty implementation, then the log and
trap messages will lose.
This PR want to add the default implementation for `env.trace` and
`env.abort` to display log and trap messages as expected.